+++ HISTORY.md
@@ -2,6 +2,10 @@
 
      - Backwards-incompatible: jsoncpp library now required
 
+     - Backwards-incompatible: transform reuse now treated as an error
+
+     - There may be a backwards-incompatible change in the transform API
+
 - Version 6:
 
      - From Masoud: New transform legendre_detrender (polynomial fitting along either time or frequency axis)